swagger-generator/requirements.txt
jsonmerge
PyYAML